Moved view into repeaters.py
I don't like having a single-view class but don't know what else to do with this one.
Consider moving it into corehq.motech.views?
