+++ corehq/apps/domain/views/__init__.py
@@ -62,9 +62,11 @@ from corehq.apps.domain.views.pro_bono import (
     ProBonoStaticView,
     ProBonoView,
 )
+'''
 from corehq.apps.domain.views.repeaters import (
     generate_repeater_payloads,
 )
+'''
 from corehq.apps.domain.views.settings import (
     BaseProjectSettingsView,
     DefaultProjectSettingsView,
